feat(WebGPU): Added environment mapping
Added environment mapping to WebGPU. The background is controlled by the
renderer's BackgroundTexture, and the strength of it's diffuse and
specular contributions can be set with BackgroundTextureDiffuseStrength
and BackgroundTextureSpecularStrength respectively (both default to
0.5). In order for surface roughness to work properly, mipmaps should be
specified for the BackgroundTexture by setting its mipLevel. Diffuse
lighting is specified by the last mipmap, so generating until a 1x1
texture is not recommended.